visualizer and audio is now instantiated once and has sources modified
commented out button ('Song') components in order to just have a single canvas and single audio element that will be modified.

Tooled around and got it so that the context, audio source, canvas, and audio are all state variables so I don't have to worry about creating additional source nodes.

Not sure the best way to handle cancelling the animation frame loop when unloading the song files.

No more errors when loading the page, so taht's cool